added test to MainActivity not null, toolbar not null and drawer not null
